Resume-App/Backend-Testing/Interview/requirements.txt
openai-whisper 
PyAudio 
python-multipart 
flask
flask_cors
dotenv
google-genai
gtts
edge_tts
asyncio
